// diploma/diploma/ClientApp/src/generated/client/model/updateUserInfoCommand.ts
/**
 * diploma
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateUserInfoCommand { 
    callerId?: string;
    firstName?: string;
    lastName?: string;
    patronymic?: string;
    additionalInfo?: string;
}